move tests to another folder
