+++ app/src/androidTest/java/com/example/moviesearchapplication/ui/LikeViewAction.kt
@@ -1,4 +1,4 @@
-package com.example.moviesearchapplication
+package com.example.moviesearchapplication.ui
 
 import android.view.View
 import androidx.test.espresso.UiController
